Made changes to notification scheme.
Idea is that there is a table with pending matches which has the message and ids to try and the attempts and statuses.

initial match query sets up table entry and sends notification to first user
When user receives match request, if they accept, status is set to accepted and notification sent back to user and then local notifications scheduled for both. If they reject, status set to rejected, notification sent back, attempts increased, user accesses table entry using ID and removes this user and trys the next one. If missed, status is missed and user does the same except does not remove the current user.

If attempts = 15 or matches is [] in the table entry, user has a local notification to try again.

So basically, all the notifications have become in a way sequential.

Need to make a router for match notifications Will need a getMatchNotification, updateMatchNotification, not sure what else.

Make a list of all the match types somewhere.